docs(html): merge the html template examples into one
Fold the per-file template-options example into the Eta example, so a single
`examples/html-template` covers templating, partial dependency capture via
`addDependency`, and per-file `template` overrides through `module.rules`.
