--- examples/html-template-per-file/build.js
@@ -1,3 +1,0 @@
-global.NO_TARGET_ARGS = true;
-global.NO_PUBLIC_PATH = true;
-require("../build-common");
