--- examples/html-template-per-file/src/index.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-	<head>
-		<meta charset="utf-8" />
-		<title><%= it.title %></title>
-	</head>
-	<body>
-		<h1><%= it.heading %></h1>
-		<p>Rendered with the default Eta options (<code>&lt;% %&gt;</code> tags).</p>
-	</body>
-</html>
