+++ examples/html-template/template.md
@@ -1,15 +1,23 @@
 This example demonstrates the experimental HTML modules support together with
-the `module.parser.html.template` option. The HTML entry is written as an
-[Eta](https://eta.js.org/) template; `template` compiles it to plain HTML
-**before** webpack parses it, so the URLs the template emits (the `<img>` source
-and the `<script src>`) are still discovered and bundled as regular webpack
-dependencies. The rewritten HTML is emitted as `dist/index.html`.
+the `module.parser.html.template` option, using [Eta](https://eta.js.org/) as
+the templating engine. `template` compiles each HTML entry to plain HTML
+**before** webpack parses it, so the URLs the template emits (the `<img>`
+source and the `<script src>`) are still discovered and bundled as regular
+webpack dependencies.
 
-The template also `include`s a partial (`src/footer.eta`). Eta resolves
-partials by reading files, so the config wraps `eta.readFile` to record every
-partial read and calls the context's `addDependency` for each — that way
-editing `footer.eta` triggers a rebuild and invalidates the persistent cache,
-even though the partial is never a webpack module itself.
+It shows three things:
+
+- **Templating** — `src/index.html` is rendered with Eta and a data object
+  (title, list items, image URL).
+- **Dependency capture** — the template `include`s a partial
+  (`src/footer.eta`). Eta resolves partials by reading files, so the config
+  wraps `eta.readFile` to record every partial read and calls the context's
+  `addDependency` for each, so editing `footer.eta` triggers a rebuild and
+  invalidates the cache even though it never becomes a webpack module.
+- **Per-file options** — `src/special.html` is matched by a `module.rules`
+  entry that hands it a differently-configured Eta (custom `{{ }}` tags,
+  `autoEscape` disabled) and its own data. `rule.parser` merges over
+  `module.parser.html`, so that `template` wins only for the matched file.
 
 # webpack.config.js
 
@@ -19,14 +27,10 @@ _{{webpack.config.js}}_
 
 # src/index.html
 
+Default Eta tags (`<%= %>`), with an `include` and a `<script src>`.
+
 ```html
 _{{src/index.html}}_
-```
-
-# src/app.js
-
-```javascript
-_{{src/app.js}}_
 ```
 
 # src/footer.eta
@@ -38,13 +42,30 @@ through `addDependency`, not bundled as a module.
 _{{src/footer.eta}}_
 ```
 
-# dist/index.html
+# src/app.js
 
-The template has been rendered (title, list items) and every URL has been
-rewritten to point at the emitted assets.
+```javascript
+_{{src/app.js}}_
+```
+
+# src/special.html
+
+Custom Eta tags (`{{= }}`) and unescaped output, selected by the rule.
+
+```html
+_{{src/special.html}}_
+```
+
+# dist/index.html
 
 ```html
 _{{dist/index.html}}_
+```
+
+# dist/special.html
+
+```html
+_{{dist/special.html}}_
 ```
 
 # Info
